Remove path.rs
Current `ObjectPath` implementation represents a sequence of string,
but this isn't fit for ItemHandler, bacause we access each items by
its id or place, not by strings.
Indeed, for mod, we have to represent ach kind of items as json. Even
in this case, 'path' implementation is not suitable beacause it's not
compatible with Rust's enum.